Basic Movement Script and State Machine
Programmed a basic state machine with three states (menu, realtime,
strategy, defaults to menu). During runtime, press "1" on the keyboard
to change to menu, "2" to change to realtime, and "3" to change to
strategy.

Set up the basics of a movement script. During runtime, while in the
"realtime" state, click (or touch, on a touch screen) on any place on
the plane. The player will move there. We can always adjust speed,
rotation, etc... later.

As of now, not too much is commented. I typically get a piece of code
running and comment it afterwards, so that'll likely be my next thing to
do tomorrow.