Rotation and Movement (mostly) Fixed
The camera now follows the player when he moves. You can rotate before,
during, and after movement with no adverse effect, save for an annoying
"feature" where the camera rotates on its own before the player starts
to move. It (hopefully) shouldn't take too long to fix, but for now,
we're able to move on. :smile: